+++ srcs/bobj/bobj_loader.h
@@ -32,12 +32,12 @@ typedef struct s_material
 typedef struct s_element
 {
 	unsigned int	*indices;
-	size_t			indices_size;
-	size_t			indices_value_amount;
-	size_t			index_amount;
-	size_t			index_value_amount;
+	int				indices_size;
+	int				indices_value_amount;
+	int				index_amount;
+	int				index_value_amount;
 
-	size_t			face_count;
+	int				face_count;
 
 	t_material		*material; // address of the material you should use;
 
@@ -67,22 +67,22 @@ typedef struct		s_mesh
 
 	// Vertices
 	float			*vertices;
-	size_t			vertices_size;
-	size_t			vertices_value_amount;
-	size_t			vertex_amount;
-	size_t			vertex_value_amount;
+	int				vertices_size;
+	int				vertices_value_amount;
+	int				vertex_amount;
+	int				vertex_value_amount;
 	// Texture Coordinates
 	float			*uvs;
-	size_t			uvs_size;
-	size_t			uvs_value_amount;
-	size_t			uv_amount;
-	size_t			uv_value_amount;
+	int				uvs_size;
+	int				uvs_value_amount;
+	int				uv_amount;
+	int				uv_value_amount;
 	// Normals
 	float			*normals;
-	size_t			normals_size;
-	size_t			normals_value_amount;
-	size_t			normal_amount;
-	size_t			normal_value_amount;
+	int				normals_size;
+	int				normals_value_amount;
+	int				normal_amount;
+	int				normal_value_amount;
 
 	t_element		*elements; // array of t_element, delimetered with usemtl in .obj;
 	int				element_amount;
@@ -102,20 +102,20 @@ struct s_obj
 	// obj_verts, obj_norms and obj_uvs, are only used in the obj_load():er;
 	//	so those should probably not be stored in the t_obj. (useless memory usage);
 	float			*obj_verts;
-	size_t			obj_vert_amount;
-	size_t			obj_vert_value_amount;
+	int				obj_vert_amount;
+	int				obj_vert_value_amount;
 
 	float			*obj_norms;
-	size_t			obj_norm_amount;
-	size_t			obj_norm_value_amount;
+	int				obj_norm_amount;
+	int				obj_norm_value_amount;
 
 	float			*obj_uvs;
-	size_t			obj_uv_amount;
-	size_t			obj_uv_value_amount;
+	int				obj_uv_amount;
+	int				obj_uv_value_amount;
 
 	// Materials
 	t_material		*materials; // array of t_materials;
-	size_t			material_amount;
+	int				material_amount;
 
 	t_mesh			*meshes; // array of meshi;
 	int				mesh_amount;
@@ -123,9 +123,9 @@ struct s_obj
 
 int					obj_load(t_obj *obj, const char *obj_file_path);
 void				obj_info(t_obj *obj);
-t_material			*material_load(size_t *mat_amount, const char *root_path, const char *mat_file);
-t_material			*get_material(t_material *materials, size_t mats, char *to_find);
-t_mesh				*get_mesh(t_mesh *meshes, size_t mesh_amount, char *name);
+t_material			*material_load(int *mat_amount, const char *root_path, const char *mat_file);
+t_material			*get_material(t_material *materials, int mats, char *to_find);
+t_mesh				*get_mesh(t_mesh *meshes, int mesh_amount, char *name);
 int					get_index_of(t_obj *obj, int v, int vn, int vt);
 
 // Print
